Update .gitignore to ignore more files
This commit updates the .gitignore file to ignore additional specific files and directories. The update ensures that numerous .idea configuration files, images under Faces and Unsorted/Sorted directories, and other miscellaneous project files are excluded from the repository. As a result, clutter is reduced and the potential of unintentionally committing unnecessary files is significantly minimized.